+++ messaging-core/src/main/java/com/nabla/sdk/core/domain/entity/MimeType.kt
@@ -1,29 +1,69 @@
 package com.nabla.sdk.core.domain.entity
 
-public sealed interface MimeType {
+public sealed class MimeType(public val type: String) {
+    public abstract val subtype: String
+
     public val stringRepresentation: String
+        get() = "$type/$subtype"
 
-    public enum class Image(override val stringRepresentation: String) : MimeType {
-        JPEG("image/jpeg"),
-        PNG("image/png"),
+    public sealed class Image(public override val subtype: String) : MimeType(TYPE) {
+        public object Jpeg : Image("jpeg")
+        public object Png : Image("png")
+        public data class Other(public override val subtype: String) : Image(subtype)
+
+        internal companion object {
+            internal const val TYPE = "image"
+
+            fun fromSubtype(subtype: String) = when (subtype) {
+                Jpeg.subtype -> Jpeg
+                Png.subtype -> Png
+                else -> Other(subtype)
+            }
+        }
     }
 
-    public enum class Application(override val stringRepresentation: String) : MimeType {
-        PDF("application/pdf"),
+    public sealed class Application(public override val subtype: String) : MimeType(TYPE) {
+        public object Pdf : Application("pdf")
+        public data class Other(public override val subtype: String) : Application(subtype)
+
+        internal companion object {
+            internal const val TYPE = "application"
+
+            fun fromSubtype(subtype: String) = when (subtype) {
+                Pdf.subtype -> Pdf
+                else -> Other(subtype)
+            }
+        }
     }
 
-    public enum class Audio(override val stringRepresentation: String) : MimeType {
-        MP3("audio/mpeg"),
+    public sealed class Audio(public override val subtype: String) : MimeType(TYPE) {
+        public object Mp3 : Audio("mp3")
+        public data class Other(public override val subtype: String) : Audio(subtype)
+
+        internal companion object {
+            internal const val TYPE = "audio"
+
+            fun fromSubtype(subtype: String) = when (subtype) {
+                Mp3.subtype -> Mp3
+                else -> Other(subtype)
+            }
+        }
     }
 
     public companion object {
         @Throws(IllegalArgumentException::class)
-        public fun fromStringRepresentation(representation: String): MimeType = when (representation) {
-            Image.JPEG.stringRepresentation -> Image.JPEG
-            Image.PNG.stringRepresentation -> Image.PNG
-            Application.PDF.stringRepresentation -> Application.PDF
-            Audio.MP3.stringRepresentation -> Audio.MP3
-            else -> throw NablaException.Internal(IllegalStateException("Unhandled mimeType: $representation"))
+        public fun fromStringRepresentation(representation: String): MimeType {
+            return try {
+                val (type, subtype) = representation.split("/")
+                when (type) {
+                    Image.TYPE -> Image.fromSubtype(subtype)
+                    Application.TYPE -> Application.fromSubtype(subtype)
+                    Audio.TYPE -> Audio.fromSubtype(subtype)
+                    else -> throw IllegalStateException("Unhandled mimeType: $representation")
+                }
+            } catch (exception: Exception) {
+                throw NablaException.Internal(exception)
+            }
         }
     }
 }