+++ messaging-ui/src/main/java/com/nabla/sdk/core/ui/helpers/mediapicker/FileNameHelper.kt
@@ -4,7 +4,7 @@ import java.text.SimpleDateFormat
 import java.util.Date
 import java.util.Locale
 
-internal fun generateFileName(extension: String): String {
+internal fun generateFileName(extension: String?): String {
     val dateFormatter = SimpleDateFormat("yyyyMMddHHmmss", Locale.getDefault())
-    return "${dateFormatter.format(Date())}.$extension"
+    return "${dateFormatter.format(Date())}${if (extension != null) ".$extension" else ""}"
 }